Rename __version__.py to _version.py
This avoids a name collision with mypackage.__version__
(the string).

In addition, the single underscore shows it's a 'private' module
that has no special meaning to the Python interpreter.
