+++ mypackage/__init__.py
@@ -1,5 +1,5 @@
 
-from .__version__ import VERSION
+from ._version import VERSION
 __version__ = '.'.join(map(str, VERSION))
 
 from .core import *
